+++ CHANGELOG.md
@@ -222,4 +222,9 @@ Developmental progress by date is stored in this file.
 - Moved the HomeScreen class from main.dart to a new file, home_screen.dart
 - Created register_or_login.dart to be used as the welcoming screen for user registration / logging in
 - Created auth_gate.dart to determine if the user should be redirected to register_or_login.dart if not logged in, else home_screen.dart
-- Created auth_services.dart for authentication functions
+- Created auth_services.dart for authentication 
+
+## 2025-09-13
+- Added fields for entering email and password for Registration / Login on the Welcome screen
+- Added functional Register / Login buttons on the Welcome screen
+- Caught error messages from failed registrations / logins and displayed them to the user